+++ db/migrate/20151022134603_create_viewers.rb
@@ -1,0 +1,13 @@
+class CreateViewers < ActiveRecord::Migration
+  def change
+    create_table :viewers do |t|
+      t.string :username
+      t.string :password_plain
+      t.string :password_digest
+
+      t.timestamps null: false
+    end
+
+    add_index :viewers, :username, unique: true
+  end
+end